refactor(sdk): move ISecretProvider to domain/ports/ subfolder
