+++ src/sdks/python/envilder/__init__.py
@@ -1,9 +1,9 @@
 from envilder.application.envilder_client import EnvilderClient
 from envilder.application.map_file_parser import MapFileParser
 from envilder.domain.envilder_options import EnvilderOptions
-from envilder.domain.i_secret_provider import ISecretProvider
 from envilder.domain.map_file_config import MapFileConfig
 from envilder.domain.parsed_map_file import ParsedMapFile
+from envilder.domain.ports.secret_provider import ISecretProvider
 from envilder.domain.secret_provider_type import SecretProviderType
 from envilder.infrastructure.aws.aws_ssm_secret_provider import (
     AwsSsmSecretProvider,
